Cosmetic rename of test app.
