+++ examples/layout/README.rst
