+++ examples/layout/layout/__main__.py
@@ -1,4 +1,4 @@
-from layout_test.app import main
+from layout.app import main
 
 if __name__ == '__main__':
     main().main_loop()
