+++ examples/layout/pyproject.toml
@@ -10,35 +10,35 @@ license = "BSD license"
 author = 'Tiberius Yak'
 author_email = "tiberius@beeware.org"
 
-[tool.briefcase.app.tayout_test]
+[tool.briefcase.app.layout]
 formal_name = "Layout Test"
 description = "A testing app"
-sources = ['tayout_test']
+sources = ['layout']
 requires = []
 
 
-[tool.briefcase.app.tayout_test.macOS]
+[tool.briefcase.app.layout.macOS]
 requires = [
     'toga-cocoa',
 ]
 
-[tool.briefcase.app.tayout_test.linux]
+[tool.briefcase.app.layout.linux]
 requires = [
     'toga-gtk',
 ]
 
-[tool.briefcase.app.tayout_test.windows]
+[tool.briefcase.app.layout.windows]
 requires = [
     'toga-winforms',
 ]
 
 # Mobile deployments
-[tool.briefcase.app.tayout_test.iOS]
+[tool.briefcase.app.layout.iOS]
 requires = [
     'toga-iOS',
 ]
 
-[tool.briefcase.app.tayout_test.android]
+[tool.briefcase.app.layout.android]
 requires = [
     'toga-android',
 ]
